# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Velora Wear API
  description: |
    RESTful API for Velora Wear - A sustainable fashion e-commerce platform.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## User Roles
    - **customer**: Regular users who can browse and purchase products
    - **admin**: Administrators with access to management endpoints
    
  version: 1.0.0
  contact:
    name: Velora Wear API Support
    email: support@velorawear.com
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.velorawear.com/api
    description: Production server
...
paths:
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create a new order
      description: Create a new order for the authenticated user
      operationId: createOrder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

    get:
      tags:
        - Orders
      summary: Get all orders
      description: Get all orders (Admin only)
      operationId: getAllOrders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /orders/myorders:
    get:
      tags:
        - Orders
      summary: Get user's orders
      description: Get all orders for the authenticated user
      operationId: getMyOrders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Get a specific order by ID (Owner or Admin only)
      operationId: getOrderById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /orders/{id}/status:
    put:
      tags:
        - Orders
      summary: Update order status
      description: Update the status of an order (Admin only)
      operationId: updateOrderStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [Processing, Shipped, Delivered, Cancelled]
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /orders/{id}/payment:
    put:
      tags:
        - Orders
      summary: Update payment status
      description: Update the payment status of an order (Admin only)
      operationId: updatePaymentStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [Pending, Paid, Failed]
      responses:
        '200':
          description: Payment status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # Review Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    Address:
      type: object
      properties:
        street:
          type: string
          example: 123 Main St
        city:
          type: string
          example: New York
        state:
          type: string
          example: NY
        zipCode:
          type: string
          example: "10001"
        country:
          type: string
          example: USA
...
    Order:
      type: object
      properties:
        _id:
          type: string
          example: 60d5ec49f1b2c72b8c8e4a1d
        user:
          type: string
          example: 60d5ec49f1b2c72b8c8e4a1b
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        paymentMethod:
          type: string
          example: Card
        paymentStatus:
          type: string
          enum: [Pending, Paid, Failed]
          example: Paid
        itemsPrice:
          type: number
          example: 90.00
        shippingPrice:
          type: number
          example: 10.00
        taxPrice:
          type: number
          example: 9.00
        totalPrice:
          type: number
          example: 109.00
        orderStatus:
          type: string
          enum: [Processing, Shipped, Delivered, Cancelled]
          example: Processing
        deliveredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required:
        - orderItems
        - shippingAddress
        - paymentMethod
        - itemsPrice
        - shippingPrice
        - taxPrice
        - totalPrice
      properties:
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        paymentMethod:
          type: string
        itemsPrice:
          type: number
        shippingPrice:
          type: number
        taxPrice:
          type: number
        totalPrice:
          type: number
    OrderItem:
      type: object
      required:
        - product
        - name
        - quantity
        - size
        - color
        - price
      properties:
        product:
          type: string
          example: 60d5ec49f1b2c72b8c8e4a1c
        name:
          type: string
          example: Classic Cotton Tee
        quantity:
          type: integer
          example: 2
        size:
          type: string
          example: M
        color:
          type: string
          example: Black
        price:
          type: number
          example: 45.00
        image:
          type: string
          example: https://example.com/image.jpg
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Error message
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              location:
                type: string
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Invalid input data
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Not authorized, no token
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Not authorized as admin
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Resource not found
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Server error